+++ Changelog.md
@@ -51,6 +51,8 @@ v2.37.0 (not published yet)
 - UI: Enhance "Frame Mismatch" error message
 - UI: Let "Frame Mismatch" ignore cutted "Copy/Mux" runs ([#1234](/../../issues/1234))
 - NVEncC: Rearrange some parameters ([#1240](/../../issues/1240))
+- NVEncC: Add "--vpp-nvvfx-artifact-reduction" parameter ([#1240](/../../issues/1240))
+- NVEncC: Add "--vpp-nvvfx-denoise" parameter ([#1240](/../../issues/1240))
 - Update tools
     - ...
 - Update AviSynth+ plugins